# classroom-backend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Classroom Backend API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/subjects": {
            "get": {
                "summary": "List subjects",
                "description": "Get all subjects with optional search filtering and pagination.",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search by subject name or code (case-insensitive)."
                    },
                    {
                        "name": "department",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Department ID."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subjects list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubjectsListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/teachers": {
            "get": {
                "summary": "List teachers",
                "responses": {
                    "200": {
                        "description": "Teachers list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Teacher"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/classes": {
            "get": {
                "summary": "List classes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Classes list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassesListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create class",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClassCreatePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created class",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Class"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Teacher": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name"
                ]
            },
            "Class": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "subjectId": {
                        "type": "integer"
                    },
                    "teacherId": {
                        "type": "integer"
                    },
                    "capacity": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "description": {
                        "type": "string"
                    },
                    "bannerImage": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "subjectId",
                    "teacherId",
                    "capacity",
                    "status"
                ]
            },
            "ClassCreatePayload": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "subjectId": {
                        "type": "integer"
                    },
                    "teacherId": {
                        "type": "integer"
                    },
                    "capacity": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "description": {
                        "type": "string"
                    },
                    "bannerImage": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "subjectId",
                    "teacherId",
                    "capacity",
                    "status",
                    "bannerImage"
                ]
            },
            "Department": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "code": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "code",
                    "name",
                    "createdAt",
                    "updatedAt"
                ]
            },
            "Subject": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "departmentId": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    }
                },
                "required": [
                    "id",
                    "departmentId",
                    "name",
                    "code",
                    "createdAt",
                    "updatedAt",
                    "department"
                ]
            },
            "Pagination": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer"
                    },
                    "limit": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    }
                },
                "required": [
                    "page",
                    "limit",
                    "total",
                    "totalPages"
                ]
            },
            "SubjectsListResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Subject"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                },
                "required": [
                    "data",
                    "pagination"
                ]
            },
            "ClassesListResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Class"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                },
                "required": [
                    "data",
                    "pagination"
                ]
            }
        }
    }
}